requirements.txt
numpy
opencv-python-headless # cv2
Pillow # PIL
tensorflow
customtkinter
scikit-learn # sklearn
matplotlib

